# 8314569: (fs) Improve normalization of UnixPath for input with trailing slashes