--- a/src/java.base/unix/classes/sun/nio/fs/UnixPath.java
+++ b/src/java.base/unix/classes/sun/nio/fs/UnixPath.java
@@ -90,7 +90,7 @@ class UnixPath implements Path {
             prevChar = c;
         }
         if (prevChar == '/') {
-            return (n == 1) ? "/"  : input.substring(0, n - 1);
+            return (n == 1) ? "/" : input.substring(0, n - 1);
         }
         return input;
     }