Avoiding `String.substring()` call in `UnixPath.normalize()` can significantly reduce normalization costs for directories on Linux:
```java
@BenchmarkMode(Mode.AverageTime)
@OutputTimeUnit(TimeUnit.NANOSECONDS)
@Warmup(time = 2, iterations = 5)
@Measurement(time = 2, iterations = 5)
@Fork(value = 4, jvmArgs = "-Xmx1g")
public class FileToPathBenchmark {

    @Benchmark
    public Path toDirectoryPath(Data data) {
        return FileSystems.getDefault().getPath(data.directoryPath);
    }

    @State(Scope.Thread)
    public static class Data {
        private String directoryPath = "/tmp/tmp/tmp/";
    }
}
```
Results:
```
baseline

Benchmark                                                             Mode  Cnt     Score     Error   Units
FileToPathBenchmark.toDirectoryPath                                   avgt   50    40.524 ±   0.940   ns/op
FileToPathBenchmark.toDirectoryPath:·gc.alloc.rate                    avgt   50  3269.167 ± 101.018  MB/sec
FileToPathBenchmark.toDirectoryPath:·gc.alloc.rate.norm               avgt   50   175.213 ±   1.200    B/op
FileToPathBenchmark.toDirectoryPath:·gc.churn.G1_Eden_Space           avgt   50  3279.401 ± 104.252  MB/sec
FileToPathBenchmark.toDirectoryPath:·gc.churn.G1_Eden_Space.norm      avgt   50   175.756 ±   1.681    B/op
FileToPathBenchmark.toDirectoryPath:·gc.churn.G1_Survivor_Space       avgt   50     0.006 ±   0.001  MB/sec
FileToPathBenchmark.toDirectoryPath:·gc.churn.G1_Survivor_Space.norm  avgt   50    ≈ 10⁻³              B/op
FileToPathBenchmark.toDirectoryPath:·gc.count                         avgt   50  1412.000            counts
FileToPathBenchmark.toDirectoryPath:·gc.time                          avgt   50  4144.000                ms

patched

Benchmark                                                             Mode  Cnt     Score     Error   Units
FileToPathBenchmark.toDirectoryPath                                   avgt   50    33.134 ±   2.313   ns/op
FileToPathBenchmark.toDirectoryPath:·gc.alloc.rate                    avgt   50  2652.997 ± 152.758  MB/sec
FileToPathBenchmark.toDirectoryPath:·gc.alloc.rate.norm               avgt   50   115.210 ±   1.960    B/op
FileToPathBenchmark.toDirectoryPath:·gc.churn.G1_Eden_Space           avgt   50  2654.258 ± 157.862  MB/sec
FileToPathBenchmark.toDirectoryPath:·gc.churn.G1_Eden_Space.norm      avgt   50   115.200 ±   1.960    B/op
FileToPathBenchmark.toDirectoryPath:·gc.churn.G1_Survivor_Space       avgt   50     0.007 ±   0.001  MB/sec
FileToPathBenchmark.toDirectoryPath:·gc.churn.G1_Survivor_Space.norm  avgt   50    ≈ 10⁻⁴              B/op
FileToPathBenchmark.toDirectoryPath:·gc.count                         avgt   50  1320.000            counts
FileToPathBenchmark.toDirectoryPath:·gc.time                          avgt   50  3939.000                ms
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8314569](https://bugs.openjdk.org/browse/JDK-8314569): (fs) Improve normalization of UnixPath for input with trailing slashes (**Enhancement** - P4)


### Reviewers
 * [Alan Bateman](https://openjdk.org/census#alanb) (@AlanBateman - **Reviewer**)
 * [Brian Burkhalter](https://openjdk.org/census#bpb) (@bplb - **Reviewer**)
 * [Roger Riggs](https://openjdk.org/census#rriggs) (@RogerRiggs - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/15342/head:pull/15342` \
`$ git checkout pull/15342`

Update a local copy of the PR: \
`$ git checkout pull/15342` \
`$ git pull https://git.openjdk.org/jdk.git pull/15342/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 15342`

View PR using the GUI difftool: \
`$ git pr show -t 15342`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/15342.diff">https://git.openjdk.org/jdk/pull/15342.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/15342#issuecomment-1683635178)