@AlanBateman 2023-08-18
> How often is the only normalization the removal of a trailing "/"? It seems quite likely.
> If that's the only normalization, then can the `normalize` call at line 94 could be replaced with a simple `substring(0, n-1)`.
> It would save the string buffer creation and copying.
> This is pretty carefully written code so it needs a thorough review.

The most common case is probably no duplicate or trailing slash, in which case this code won't be executed. The next most common is probably the trailing slash case, in which I think you have a good point, just need to be careful not to break the "/" case.