@AlanBateman 2023-08-19
> @AlanBateman isn't "/" case handled in lines 108-109?

I think Roger's suggestion means that the trailing case would be handled in normalizeAndCheck. There are many ways to do this, we just need to keep it maintainable.