@AlanBateman 2023-08-29
@stsypanov The title on the PR is currently:

"(fs) Improve normalization of UnixPath for directories(fs) Improve normalization of UnixPath for input with trailing slashes #15342"

Something got fat fingered somewhere. Can you fix it to match the JBS issue  "(fs) Improve normalization of UnixPath for input with trailing slashes" and then we can sponsor?